+++ states_rebuilder_package/lib/src/assertions.dart
@@ -177,4 +177,15 @@ This is not allowed, because setState method of a reactive model injected using 
 
             ''';
   }
+
+  static injectingAnInjectedModel(String name) {
+    return '''
+***Injecting an already injected model***
+You are injecting the [$name] model which is already injected.
+This is not allowed.
+If we are sure you want to inject it again, try injecting it with custom name.
+
+If you are testing the widget set [Injector.enableTestMode] to true
+    ''';
+  }
 }